refactor/parsec: rename meta_elections.rs to meta_election.rs
